+++ tests/Core/Content/App/Action/AppActionTest.php
@@ -3,6 +3,7 @@
 namespace Swag\SaasConnect\Test\Core\Content\App\Action;
 
 use PHPUnit\Framework\TestCase;
+use Shopware\Core\Framework\Util\Random;
 use Shopware\Core\Framework\Uuid\Uuid;
 use Swag\SaasConnect\Core\Content\App\Action\AppAction;
 use Swag\SaasConnect\Core\Content\App\Exception\InvalidArgumentException;
@@ -16,6 +17,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         $result = new AppAction(
             $targetUrl,
@@ -26,7 +28,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
 
         $expected = [
@@ -35,6 +38,7 @@ class AppActionTest extends TestCase
                 'appVersion' => $appVersion,
                 'apiKey' => 'i am not secret',
                 'secretKey' => 'i am secret',
+                'shopId' => $shopId,
             ],
             'data' => [
                 'ids' => $ids,
@@ -54,6 +58,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -64,7 +69,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -76,6 +82,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -86,7 +93,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -98,6 +106,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -108,7 +117,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -120,6 +130,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = '';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -130,7 +141,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -142,6 +154,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = '';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -152,7 +165,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -164,6 +178,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex(), 'test'];
         new AppAction(
             $targetUrl,
@@ -174,7 +189,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -186,6 +202,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -196,7 +213,8 @@ class AppActionTest extends TestCase
             $ids,
             '',
             'i am secret',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -208,6 +226,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -218,7 +237,8 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             '',
-            's3cr3t'
+            's3cr3t',
+            $shopId
         );
     }
 
@@ -230,6 +250,7 @@ class AppActionTest extends TestCase
         $appVersion = '1.0.0';
         $entity = 'product';
         $action = 'detail';
+        $shopId = Random::getAlphanumericString(12);
         $ids = [Uuid::randomHex()];
         new AppAction(
             $targetUrl,
@@ -240,7 +261,32 @@ class AppActionTest extends TestCase
             $ids,
             'i am not secret',
             'i am secret',
-            ''
+            '',
+            $shopId
+        );
+    }
+
+    public function testInvalidShopId(): void
+    {
+        static::expectException(InvalidArgumentException::class);
+        $targetUrl = 'https://my-server.com/action';
+        $shopUrl = 'https://my-shop.com';
+        $appVersion = '1.0.0';
+        $entity = 'product';
+        $action = 'detail';
+        $shopId = '';
+        $ids = [Uuid::randomHex()];
+        new AppAction(
+            $targetUrl,
+            $shopUrl,
+            $appVersion,
+            $entity,
+            $action,
+            $ids,
+            'i am not secret',
+            'i am secret',
+            's3cr3t',
+            $shopId
         );
     }
 }